sdk-c: header reorganisation
-reorganised header so that private header are in include/internal
/!\ for now C sdk does have proper public header

-renamed kuzzlesdk.h to kuzzle_structs.h as it is not he entry point
of what could be the C SDK

-now the C SDK file structure is built during build (`make all`)
in build kuzzle-c-sdk
-> C++ sdk shall point to this folder when building...
