+++ Makefile
@@ -33,28 +33,28 @@ endif
 SDKGOPATH = go$(PATHSEP)src$(PATHSEP)github.com$(PATHSEP)kuzzleio$(PATHSEP)sdk-go
 CGOPATH = cgo$(PATHSEP)kuzzle
 PATHSEP = $(strip $(SEP))
-ROOTOUTDIR = $(ROOT_DIR)build
+BUILD_DIR = $(ROOT_DIR)build
 GOFLAGS ?= -buildmode=c-archive
 GOFLAGSSHARED = -buildmode=c-shared
 GOSRC = .$(PATHSEP)cgo$(PATHSEP)kuzzle$(PATHSEP)
-GOTARGET = $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB)
-GOTARGETSO = $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB)
+GOTARGET = $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB)
+GOTARGETSO = $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB)
 
 export GOPATH = $(ROOT_DIR)go
 
 CORE_SRC = $(wildcard $(GOSRC)*.go)
-all:  $(ROOTOUTDIR)/c 
+all:  $(BUILD_DIR)/sdk
 
-$(ROOTOUTDIR)/pre_core:
+$(BUILD_DIR)/pre_core:
 	cd $(SDKGOPATH) && go get .$(PATHSEP)...
 	@touch $@
 
-$(ROOTOUTDIR)/core: $(CORE_SRC) Makefile
+$(BUILD_DIR)/core: $(CORE_SRC) Makefile
 ifneq ($(OS),Windows_NT)
 ifeq ($(wildcard $(GOCC)),)
 	$(error "Unable to find go compiler in $(GOCC)")
 endif
-	cd $(ROOTOUTDIR) && rm -f $(GOTARGET).* $(GOTARGETSO).*
+	cd $(BUILD_DIR) && rm -f $(GOTARGET).* $(GOTARGETSO).*
 endif
 ifeq ($(GOOS), android)
 	$(GOCC) build -o $(GOTARGET) $(GOFLAGSSHARED) $(GOSRC)
@@ -63,33 +63,38 @@ else
 endif
 	$(GOCC) build -o $(GOTARGETSO) $(GOFLAGSSHARED) $(GOSRC)
 
-	cd $(ROOTOUTDIR) && mv $(GOTARGET) $(GOTARGET).$(VERSION) && mv $(GOTARGETSO) $(GOTARGETSO).$(VERSION)
+	cd $(BUILD_DIR) && mv $(GOTARGET) $(GOTARGET).$(VERSION) && mv $(GOTARGETSO) $(GOTARGETSO).$(VERSION)
 
 ifeq ($(OS),Windows_NT)
-	$(MV) $(subst /,\,$(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk.h) kuzzle.h
+	$(MV) $(subst /,\,$(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk.h) kuzzle.h
 else
-	$(MV) $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk.h $(ROOTOUTDIR)$(PATHSEP)kuzzle.h
+	$(MV) $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk.h $(BUILD_DIR)$(PATHSEP)kuzzle.h
 endif
 	 @touch $@
 
-$(ROOTOUTDIR):
+$(BUILD_DIR):
 ifeq ($(OS),Windows_NT)
-	@if not exist $(subst /,\,$(ROOTOUTDIR)) mkdir $(subst /,\,$(ROOTOUTDIR))
+	@if not exist $(subst /,\,$(BUILD_DIR)) mkdir $(subst /,\,$(BUILD_DIR))
 else
 	mkdir -p $@
 endif
 
-$(ROOTOUTDIR)/c: $(ROOTOUTDIR) $(ROOTOUTDIR)/pre_core $(ROOTOUTDIR)/core
-	cd $(ROOTOUTDIR) && ln -srf $(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(STATICLIB)
-	cd $(ROOTOUTDIR) && ln -srf $(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(DYNLIB)
+$(BUILD_DIR)/libs: $(BUILD_DIR) $(BUILD_DIR)/pre_core $(BUILD_DIR)/core
+	cd $(BUILD_DIR) && ln -srf $(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(STATICLIB)
+	cd $(BUILD_DIR) && ln -srf $(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(DYNLIB)
 	@touch $@
 
-package: $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION)
-	mkdir -p $(ROOTOUTDIR)$(PATHSEP)$(SDK_FOLDER_NAME)/lib
-	mkdir -p $(ROOTOUTDIR)$(PATHSEP)$(SDK_FOLDER_NAME)/include
-	cp -fr $(ROOT_DIR)$(PATHSEP)include$(PATHSEP)*.h $(ROOTOUTDIR)$(PATHSEP)kuzzle-cpp-sdk/include
-	cp $(ROOTOUTDIR)$(PATHSEP)*.a $(ROOTOUTDIR)$(PATHSEP)*.so  $(ROOTOUTDIR)$(PATHSEP)$(SDK_FOLDER_NAME)/lib
-	mkdir -p deploy && cd $(ROOTOUTDIR) && tar cfz ..$(PATHSEP)deploy$(PATHSEP)kuzzlesdk-c-$(VERSION)-$(ARCH).tar.gz $(SDK_FOLDER_NAME)
+$(BUILD_DIR)/sdk: $(BUILD_DIR)/libs
+	mkdir -p $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/lib
+	mkdir -p $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/include/internal
+	cp $(BUILD_DIR)$(PATHSEP)kuzzle.h  $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/include/internal
+	cp -fr $(ROOT_DIR)$(PATHSEP)include $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)
+	cp -fr $(ROOT_DIR)$(PATHSEP)include $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)
+	cp $(BUILD_DIR)$(PATHSEP)*.a* $(BUILD_DIR)$(PATHSEP)*.so*  $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/lib
+	@touch $@
+
+package: $(BUILD_DIR)/sdk
+	mkdir -p deploy && cd $(BUILD_DIR) && tar cfz ..$(PATHSEP)deploy$(PATHSEP)kuzzlesdk-c-$(VERSION)-$(ARCH).tar.gz $(SDK_FOLDER_NAME)
 
 clean:
 ifeq ($(OS),Windows_NT)
