+++ cgo/kuzzle/auth.go
@@ -19,8 +19,8 @@ package main
 
   #include <stdlib.h>
   #include <errno.h>
-  #include "kuzzlesdk.h"
-  #include "sdk_wrappers_internal.h"
+  #include "internal/kuzzle_structs.h"
+  #include "internal/sdk_wrappers_internal.h"
 */
 import "C"
 import (
