+++ cgo/kuzzle/collection.go
@@ -16,8 +16,8 @@ package main
 
 /*
 	#cgo CFLAGS: -I../../include
-	#include "kuzzlesdk.h"
-	#include "sdk_wrappers_internal.h"
+	#include "internal/kuzzle_structs.h"
+	#include "internal/sdk_wrappers_internal.h"
 */
 import "C"
 import (
