+++ cgo/kuzzle/errors.go
@@ -17,7 +17,7 @@ package main
 /*
 #cgo CFLAGS: -I../../include
 
-#include "kuzzlesdk.h"
+#include "internal/kuzzle_structs.h"
 */
 import "C"
