+++ cgo/kuzzle/memory_storage.go
@@ -19,8 +19,8 @@ package main
 
   #include <stdlib.h>
   #include <string.h>
-  #include "kuzzlesdk.h"
-  #include "sdk_wrappers_internal.h"
+  #include "internal/kuzzle_structs.h"
+  #include "internal/sdk_wrappers_internal.h"
 
   static void assign_geopos(double (*ptr)[2], int idx, double lon, double lat) {
     ptr[idx][0] = lon;
